refactor: server actions, cache and misc
