+++ app/(connected)/(with-sidebar)/(with-chat)/pilotage/page.tsx
@@ -10,12 +10,10 @@ import {
 } from 'components/PageNavigationPortals'
 import { estUserMilo, peutAccederAuxSessions } from 'interfaces/conseiller'
 import { AnimationCollectivePilotage } from 'interfaces/evenement'
-import {
-  getActionsAQualifierServerSide,
-  getSituationsNonProfessionnelles,
-} from 'services/actions.service'
+import { getActionsAQualifierServerSide } from 'services/actions.service'
 import { getConseillerServerSide } from 'services/conseiller.service'
 import { getAnimationsCollectivesACloreServerSide } from 'services/evenements.service'
+import { getSituationsNonProfessionnelles } from 'services/referentiel.service'
 import {
   getSessionsACloreServerSide,
   SessionsAClore,
