+++ app/(connected)/(with-sidebar)/(without-chat)/mes-jeunes/[idJeune]/actions/[idAction]/modification/page.tsx
@@ -9,11 +9,11 @@ import {
 } from 'components/PageNavigationPortals'
 import { StatutAction } from 'interfaces/action'
 import { estUserMilo } from 'interfaces/conseiller'
+import { getAction } from 'services/actions.service'
 import {
-  getAction,
+  getActionsPredefinies,
   getSituationsNonProfessionnelles,
-} from 'services/actions.service'
-import { getActionsPredefinies } from 'services/referentiel.service'
+} from 'services/referentiel.service'
 import { getMandatorySessionServerSide } from 'utils/auth/auth'
 
 type ModificationActionParams = { idAction: string }
