+++ app/(connected)/(with-sidebar)/(without-chat)/mes-jeunes/[idJeune]/actions/nouvelle-action/page.tsx
@@ -6,9 +6,11 @@ import {
   PageHeaderPortal,
   PageRetourPortal,
 } from 'components/PageNavigationPortals'
-import { getSituationsNonProfessionnelles } from 'services/actions.service'
 import { getIdentitesBeneficiairesServerSide } from 'services/beneficiaires.service'
-import { getActionsPredefinies } from 'services/referentiel.service'
+import {
+  getActionsPredefinies,
+  getSituationsNonProfessionnelles,
+} from 'services/referentiel.service'
 import { getMandatorySessionServerSide } from 'utils/auth/auth'
 
 type NouvelleActionParams = { idJeune: string }
