+++ components/chat/ListeConversations.tsx
@@ -91,8 +91,7 @@ function ListeConversations(
 
       {conversations && conversations.length > 0 && (
         <ButtonLink
-          // FIXME : dirty fix, problème de rafraichissement des listes de diffusion
-          href={'/mes-jeunes/envoi-message-groupe?misc=' + Math.random()}
+          href='/mes-jeunes/envoi-message-groupe'
           style={ButtonStyle.PRIMARY}
           className='absolute bottom-8 self-center'
         >
