+++ server-actions/actions.server-actions.ts
@@ -1,9 +1,12 @@
 'use server'
 
-import { revalidateTag } from 'next/cache'
-
-import { apiPut } from 'clients/api.client'
-import { ActionFormData, actionStatusToJson } from 'interfaces/json/action'
+import { apiPost, apiPut } from 'clients/api.client'
+import {
+  ActionFormData,
+  actionStatusToJson,
+  QualificationActionJson,
+} from 'interfaces/json/action'
+import { ACTION_CACHE_TAG } from 'services/actions.service'
 import { getSessionServerSide } from 'utils/auth/auth'
 
 export async function modifierAction(
@@ -23,6 +26,38 @@ export async function modifierAction(
     codeQualification: modifications.codeCategorie,
   }
 
-  await apiPut(`/actions/${idAction}`, actionModifiee, session!.accessToken)
-  revalidateTag('action-' + idAction)
+  await apiPut(
+    `/actions/${idAction}`,
+    actionModifiee,
+    session!.accessToken,
+    Object.values(ACTION_CACHE_TAG)
+  )
+}
+
+export async function qualifier(
+  idAction: string,
+  type: string,
+  options?: {
+    dateFinModifiee?: string
+    commentaire?: string
+  }
+) {
+  const session = await getSessionServerSide()
+
+  const payload: {
+    codeQualification: string
+    dateFinReelle?: string
+    commentaireQualification?: string
+  } = { codeQualification: type }
+
+  if (options?.dateFinModifiee) payload.dateFinReelle = options.dateFinModifiee
+  if (options?.commentaire)
+    payload.commentaireQualification = options.commentaire
+
+  await apiPost<QualificationActionJson>(
+    `/actions/${idAction}/qualifier`,
+    payload,
+    session!.accessToken,
+    Object.values(ACTION_CACHE_TAG)
+  )
 }
