+++ services/favoris.service.ts
@@ -1,11 +1,17 @@
 import { apiGet } from 'clients/api.client'
+import { MetadonneesFavoris } from 'interfaces/beneficiaire'
 import { Offre, Recherche } from 'interfaces/favoris'
+import {
+  jsonToMetadonneesFavoris,
+  MetadonneesFavorisJson,
+} from 'interfaces/json/beneficiaire'
 import {
   jsonToOffre,
   jsonToRecherche,
   OffreJson,
   RechercheJson,
 } from 'interfaces/json/favoris'
+import { ApiError } from 'utils/httpClient'
 
 export async function getOffres(
   idJeune: string,
@@ -13,7 +19,8 @@ export async function getOffres(
 ): Promise<Offre[]> {
   const { content: offresJson } = await apiGet<OffreJson[]>(
     `/jeunes/${idJeune}/favoris`,
-    accessToken
+    accessToken,
+    'favoris'
   )
   return offresJson.map(jsonToOffre)
 }
@@ -24,8 +31,27 @@ export async function getRecherchesSauvegardees(
 ): Promise<Recherche[]> {
   const { content: recherchesJson } = await apiGet<RechercheJson[]>(
     `/jeunes/${idJeune}/recherches`,
-    accessToken
+    accessToken,
+    'favoris'
   )
 
   return recherchesJson.map(jsonToRecherche)
+}
+
+export async function getMetadonneesFavorisJeune(
+  idJeune: string,
+  accessToken: string
+): Promise<MetadonneesFavoris | undefined> {
+  try {
+    const { content: metadonneesFavoris } = await apiGet<{
+      favoris: MetadonneesFavorisJson
+    }>(`/jeunes/${idJeune}/favoris/metadonnees`, accessToken, 'favoris')
+    return jsonToMetadonneesFavoris(metadonneesFavoris)
+  } catch (e) {
+    if (e instanceof ApiError && e.statusCode === 404) {
+      return undefined
+    }
+
+    throw e
+  }
 }
