+++ services/referentiel.service.ts
@@ -1,8 +1,14 @@
 import { getSession } from 'next-auth/react'
 
 import { apiGet } from 'clients/api.client'
-import { ActionPredefinie } from 'interfaces/action'
+import {
+  ActionPredefinie,
+  SituationNonProfessionnelle,
+} from 'interfaces/action'
+import { CODE_QUALIFICATION_NON_SNP } from 'interfaces/json/action'
 import { Agence, Commune, Localite, Metier } from 'interfaces/referentiel'
+
+const TAG = 'referentiel'
 
 export function getAgencesServerSide(
   structure: string,
@@ -33,7 +39,8 @@ export async function getMetiers(query: string): Promise<Metier[]> {
   const session = await getSession()
   const { content: metiers } = await apiGet<Metier[]>(
     `/referentiels/metiers?recherche=${encodeURIComponent(query)}`,
-    session!.accessToken
+    session!.accessToken,
+    TAG
   )
   return metiers
 }
@@ -43,9 +50,26 @@ export async function getActionsPredefinies(
 ): Promise<ActionPredefinie[]> {
   const { content: actionsPredefinies } = await apiGet<ActionPredefinie[]>(
     `/referentiels/actions-predefinies`,
-    accessToken
+    accessToken,
+    TAG
   )
   return actionsPredefinies
+}
+
+export async function getSituationsNonProfessionnelles(
+  { avecNonSNP }: { avecNonSNP: boolean },
+  accessToken: string
+): Promise<SituationNonProfessionnelle[]> {
+  const { content } = await apiGet<SituationNonProfessionnelle[]>(
+    '/referentiels/qualifications-actions/types',
+    accessToken,
+    TAG
+  )
+  return avecNonSNP
+    ? content
+    : content.filter(
+        (categorie) => categorie.code !== CODE_QUALIFICATION_NON_SNP
+      )
 }
 
 async function getAgences(
@@ -54,7 +78,8 @@ async function getAgences(
 ): Promise<Agence[]> {
   const { content: agences } = await apiGet<Agence[]>(
     `/referentiels/agences?structure=${structure}`,
-    accessToken
+    accessToken,
+    TAG
   )
   return agences
 }
@@ -63,7 +88,8 @@ async function getLocalites(path: string, query: string): Promise<Localite[]> {
   const session = await getSession()
   const { content: localites } = await apiGet<Localite[]>(
     path + `recherche=${encodeURIComponent(query)}`,
-    session!.accessToken
+    session!.accessToken,
+    TAG
   )
 
   return Array.from(
